- Waypoint tracks shouldn't go back to an earlier position as it causes an assertion failure (to be fixed later) so I set them to not end at the origin.  - Added Gabe's heading controller on top of the L2+ controller in my controller in the boat_sim to fix following the L2 vector.  - Fixed an unpacking error in the throttle command in the Plant in Boat_sim.mdl.  - Fixed the ballast commands to be 0.  - Test runs now look pretty stable.

+++ Code/basic_model/PlotTrack.m
@@ -21,13 +21,13 @@ myLegend{end + 1} = 'Test waypoints';
 % reset to properly illustrate where the boat thought the waypoints were.
 offsets = position(nextTrack > 0,:);
 if size(offsets,1) > 0
-    figure8_waypoints_offset = figure8_waypoints + repmat(offsets(1,:),size(figure8_waypoints,1),1);
+    figure8_waypoints_offset = double(figure8_waypoints) + repmat(offsets(1,:),size(figure8_waypoints,1),1);
     plot(figure8_waypoints_offset(:,2)', figure8_waypoints_offset(:,1)', 'g^', 'MarkerSize', 10);
     text(figure8_waypoints_offset(:,2)'+10, figure8_waypoints_offset(:,1)'-10, cellstr({int2str((1:size(figure8_waypoints_offset,1))')}), 'Color', 'g');
     myLegend{end + 1} = 'Figure-8 waypoints';
 end
 if size(offsets,1) > 1
-    sampling_waypoints_offset = sampling_waypoints + repmat(offsets(2,:),size(sampling_waypoints,1),1);
+    sampling_waypoints_offset = double(sampling_waypoints) + repmat(offsets(2,:),size(sampling_waypoints,1),1);
     plot(sampling_waypoints_offset(:,2)', sampling_waypoints_offset(:,1)', 'b^', 'MarkerSize', 10);
     text(sampling_waypoints_offset(:,2)'+10, sampling_waypoints_offset(:,1)'-10, cellstr({int2str((1:size(sampling_waypoints_offset,1))')}), 'Color', 'b');
     myLegend{end + 1} = 'Search waypoints';
